requirements.txt
flask
opentelemetry-instrumentation
opentelemetry-exporter-jaeger
opentelemetry-instrumentation-flask
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-requests